requirements.txt
openai>=1.40.0
tqdm>=4.66.0
pandas>=2.1.0
numpy>=1.26.0
python-dotenv>=1.0.0
jsonlines>=4.0.0
jinja2>=3.1.0
